Close Rack::BodyProxy after handling it
When using a ThreadPool threads are reused, this had the effect that
Rack::Lock middleware tried to lock twice on the same mutex.
The lock happens here:
https://github.com/rack/rack/blob/master/lib/rack/lock.rb#L13

The ruby documentation regarding to Mutex#lock confirms this:
http://ruby-doc.org/core-2.2.0/Mutex.html#lock-method

This was no problem when using Thread.new because on every request a new
Thread was spawned.

When calling BodyProxy#close the mutex is released.
https://github.com/rack/rack/blob/master/lib/rack/lock.rb#L16 the mutex
unlock is passed in as block.

https://github.com/rack/rack/blob/master/lib/rack/body_proxy.rb#L23
mutex unlock is called when body proxy is closed.